requirements.txt
flask
flask-cors
python-dotenv
Flask-SQLAlchemy
SQLAlchemy
PyJWT
mysqlclient
